Modify describe and it block in userCanListofProducts.feature.js
