+++ package.json
@@ -17,7 +17,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "start:silent": "PORT=3001 BROWSER=none react-scripts start",
-    "cy:open": "cypress open",
+    "cy:open": "yarn start --silent & cypress open",
     "cypress": "start-server-and-test start:silent http://localhost:3001/ cy:open"
   },
   "eslintConfig": {
