+++ src/App.jsx
@@ -1,11 +1,13 @@
-import React from 'react';
+import React from "react";
+import DisplayProductData from "./components/DisplayProductData";
 
 const App = () => {
   return (
     <>
       <h1>Slowfood</h1>
+      <DisplayProductData />
     </>
   );
-}
+};
 
 export default App;
