+++ src/index.js
@@ -2,6 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
+import axios from "axios";
+
+axios.defaults.baseURL = "http://localhost:3000/api/";
 
 ReactDOM.render(<App />, document.getElementById('root'));
